+++ CHANGES.rst
@@ -13,8 +13,12 @@ Changelog
   nonzero values are reported when the credible region is smaller than one
   pixel.
 
-- The Earth plot marker is now unfilled by default. It is no longer necessary
-  to specify ``markerfacecolor='none'``.
+- Some improvements to plot markers:
+
+  - The Earth plot marker is now unfilled by default. It is no longer necessary
+    to specify ``markerfacecolor='none'``.
+
+  - Added Sun symbol marker (circle with dot).
 
 - Several improvements in ``ligo-skymap-from-samples``, including a speedup
   from about 1 min to 20 s for an input of 10,000 samples on high core count